Made some updates to server and app.js
Tried to make a server but did not work. Updated some other things as well. Fixed the bookmark section on home view so now it outputs anime that is bookmarked. I also fixed the bookmarkView section so it also shows the bookmarks. I made it so that when you unbookmark something it automatically leaves bookmark sections and vice versa when bookmarked they automatically show up in bookmark sections.