Fixed up the server.js plus other fixes to app.js
Fixed the server so that is working now.

I also fixed the search function so it now searches for multiple searches instead of one.

Made another fetch for when you click on the anime. It now goes to a different view with anime information. Added that functionality to all the anime buttons

cleaned up the files that I did not need