Updates to alot of things
I added login/logout functionality with Anilist API

updated server.js to work with new login functionality. Also Added the bookmark pouchDB to the server instead of app.js. Made it so that the fetches that don't need authorization still work

Fixed a bunch of the problems with bookmarking and unbookmarking. Bookmark UI is consistent now on reload

Only allowed to view bookmarks when signed in but still able to see the home page fetches except for bookmarks

Fixed the search function.

added functionality to all the view all buttons. Takes you to a new view that has all the 25 cards that were previously fetched

Changed the buttons for login/logout in the index.html. Also added the different views for the view all buttons. Deleted the old away to login/logout