Updates to index.html,server.js and app.js
In index.html I added a new selection for filters in the bookmark section. I have also added it to the account view so it finally displays something.

In server.js I added the status selections to server instead of being in app.js

In app.js fixed anime status so now it works and is persistent. Also added a filter function so that you can filter bookmarks status to see all of one status. Added the values to the account page.